[refactor/#43] SubmittedFile 클래스의 위치 변경
