+++ src/main/kotlin/goodspace/teaming/global/entity/file/SubmittedFile.kt
@@ -1,7 +1,7 @@
-package goodspace.teaming.global.entity.aissgnment
+package goodspace.teaming.global.entity.file
 
 import goodspace.teaming.global.entity.BaseEntity
-import goodspace.teaming.global.entity.file.File
+import goodspace.teaming.global.entity.aissgnment.Submission
 import jakarta.persistence.*
 import jakarta.persistence.GenerationType.*
 import org.hibernate.annotations.SQLDelete
